Los usuarios tutorados pueden ver unicamente las salas en las que está añadido, además,pueden ver las misiosnes dentro de dichas salas
