+++ lib/Servicios/Autenticacion/Autenticacion.dart
@@ -14,20 +14,6 @@ class Autenticar{
     User? user;
 
 
-    // Trigger the authentication flow
-    /*
-    final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
-
-    // Obtain the auth details from the request
-    final GoogleSignInAuthentication? googleAuth = await googleUser
-        ?.authentication;
-
-    // Create a new credential
-    final credential = GoogleAuthProvider.credential(
-      accessToken: googleAuth?.accessToken,
-      idToken: googleAuth?.idToken,
-    );*/
-
     try {
       UserCredential userCredential = await autenticador.signInWithCredential(
           credential);
