+++ lib/Servicios/Solicitudes/AdminSolicitudes.dart
@@ -1,34 +1,44 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:retos_proyecto/vista_tutor/TabPages/pages/UsuariosTutorados.dart';
 
 class Solicitudes {
+
 //Apceptar solicitud-------------------------------------------------------------------------------------
   static aceptarSolicitud(
       String id_emisor,
       String id_sala,
       CollectionReference collectionReferenceUers,
-      User? currentUser,
+      String? idCurrentUser,
       BuildContext context) async {
+
+
     await collectionReferenceUers
-        .doc(currentUser?.uid)
+        .doc(idCurrentUser)
         .collection('rolTutorado')
         .doc(id_emisor.trim())
-        .update({
-      'salas_id': FieldValue.arrayUnion([id_sala])
-    }).then((value) async => {
+        .get()
+        .then((value) => {
+              value.exists
+                  ? value.reference.update({
+                      'salas_id': FieldValue.arrayUnion([id_sala])
+                    })
+                  : value.reference.set({
+                      'salas_id': FieldValue.arrayUnion([id_sala])
+                    })
+            })
+        .then((value) async => {
               await collectionReferenceUers
                   .doc(id_emisor)
                   .collection('rolTutor')
                   .doc(id_sala)
                   .collection('usersTutorados')
-                  .doc(currentUser?.uid.trim())
+                  .doc(idCurrentUser)
                   .set({'puntosTotal': 0}).then((value) async => {
                         await eliminarNotificacion(
                             id_sala,
                             collectionReferenceUers,
-                            currentUser,
+                            idCurrentUser,
                             context,
                             'aceptada')
                       })
@@ -39,14 +49,16 @@ class Solicitudes {
   static eliminarNotificacion(
       String id_sala,
       CollectionReference collectionReferenceUers,
-      User? currentUser,
+      String? idCurrentUser,
       BuildContext context,
       String accion) async {
+
     var colorSnackBar = accion == 'aceptada' ? Colors.green : Colors.red;
+
     await collectionReferenceUers
-        .doc(currentUser?.uid)
+        .doc(idCurrentUser)
         .collection('notificaciones')
-        .doc(currentUser?.uid)
+        .doc(idCurrentUser)
         .collection('solicitudesRecibidas')
         .doc(id_sala)
         .delete()
@@ -60,8 +72,11 @@ class Solicitudes {
   //Enviar solicitud-----------------------------------------------------------------------------------------
   static Future<bool> enviarSolicitud(String userName,
       CollectionReference collectionReferenceUsers, String idSala) async {
+
     var resultadoFinal = false;
-    FirebaseAuth auth = FirebaseAuth.instance;
+
+    User? currentUser = FirebaseAuth.instance.currentUser;
+
     await collectionReferenceUsers
         .where('nombre_usuario', isEqualTo: userName)
         .get()
@@ -74,9 +89,9 @@ class Solicitudes {
                       .collection('solicitudesRecibidas')
                       .doc(idSala)
                       .set({
-                    'id_emisor': auth.currentUser?.uid,
+                    'id_emisor': currentUser?.uid.trim(),
                     'id_sala': idSala,
-                    'nombre_emisor': auth.currentUser?.displayName
+                    'nombre_emisor': currentUser?.displayName
                   }).then((value) => {resultadoFinal = true}),
                 }
               else
@@ -85,5 +100,4 @@ class Solicitudes {
 
     return resultadoFinal;
   }
-
 }
