--- lib/vista_tutorado/Salas/ListaMisionesVtutorado.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ListMisiones_Tutorado{
-  static Widget getLista(){
-    return Text("Misiones");
-  }
-}
