--- lib/vista_tutorado/Salas/SalasAsignadas.dart
@@ -1,3 +1,0 @@
-class SalasAsignadas{
-
-}
